+++ src/term.rs
@@ -79,65 +79,100 @@ pub fn colour(colour_code: &str) -> String {
         COLOURS_ENABLED.store(true, Ordering::SeqCst);
     }
 
-    if colour_code.starts_with("#") {
-        if colour_code.len() == 7 {
-            return format!(
-                "\x1b[38;2;{};{};{}m",
-                i16::from_str_radix(colour_code.get(1..3).unwrap(), 16).unwrap(),
-                i16::from_str_radix(colour_code.get(3..5).unwrap(), 16).unwrap(),
-                i16::from_str_radix(colour_code.get(5..7).unwrap(), 16).unwrap()
-            );
+    let mut color = colour_code.to_uppercase();
+    let mut code = "\x1b[".to_string();
+
+    let bg = if color.contains("ON BLACK") {
+        color = color.replace("ON BLACK", "");
+        ";40"
+    } else if color.contains("ON RED") {
+        color = color.replace("ON RED", "");
+        ";41"
+    } else if color.contains("ON GREEN") {
+        color = color.replace("ON GREEN", "");
+        ";42"
+    } else if color.contains("ON YELLOW") {
+        color = color.replace("ON YELLOW", "");
+        ";43"
+    } else if color.contains("ON BLUE") {
+        color = color.replace("ON BLUE", "");
+        ";44"
+    } else if color.contains("ON MAGENTA") {
+        color = color.replace("ON MAGENTA", "");
+        ";45"
+    } else if color.contains("ON CYAN") {
+        color = color.replace("ON CYAN", "");
+        ";46"
+    } else if color.contains("ON WHITE") {
+        color = color.replace("ON WHITE", "");
+        ";47"
+    } else {
+        ""
+    };
+
+    // rgb(175,0,255)
+    if let Some(hex_code) = colour_code.find("#") {
+        code += &format!(
+            "38;2;{};{};{}",
+            i16::from_str_radix(&colour_code[(hex_code + 1)..(hex_code + 3)], 16)
+                .unwrap(),
+            i16::from_str_radix(&colour_code[(hex_code + 3)..(hex_code + 5)], 16)
+                .unwrap(),
+            i16::from_str_radix(&colour_code[(hex_code + 5)..(hex_code + 7)], 16)
+                .unwrap()
+        );
+    } else {
+        if color.contains("BRIGHT") {
+            if color.contains("BLACK") {
+                code += "90";
+            } else if color.contains("RED") {
+                code += "91";
+            } else if color.contains("GREEN") {
+                code += "92";
+            } else if color.contains("YELLOW") {
+                code += "93";
+            } else if color.contains("BLUE") {
+                code += "94";
+            } else if color.contains("MAGENTA") {
+                code += "95";
+            } else if color.contains("CYAN") {
+                code += "96";
+            } else if color.contains("WHITE") {
+                code += "97";
+            } else {
+                return "".to_string();
+            }
         } else {
-            panic!(
-                "Unknown hex colour format [{}]; Valid type: [hex (#00ff00)]",
-                colour_code
-            );
+            if color.contains("BLACK") {
+                code += "30";
+            } else if color.contains("RED") {
+                code += "31";
+            } else if color.contains("GREEN") {
+                code += "32";
+            } else if color.contains("YELLOW") {
+                code += "33";
+            } else if color.contains("BLUE") {
+                code += "34";
+            } else if color.contains("MAGENTA") {
+                code += "35";
+            } else if color.contains("CYAN") {
+                code += "36";
+            } else if color.contains("WHITE") {
+                code += "37";
+            } else {
+                return "".to_string();
+            }
         }
     }
 
-    let color = colour_code.to_uppercase();
-    let mut code = "\x1b[".to_string();
-
-    if color.contains("BLACK") {
-        code += "30";
-    } else if color.contains("RED") {
-        code += "31";
-    } else if color.contains("GREEN") {
-        code += "32";
-    } else if color.contains("YELLOW") {
-        code += "33";
-    } else if color.contains("BLUE") {
-        code += "34";
-    } else if color.contains("MAGENTA") {
-        code += "35";
-    } else if color.contains("CYAN") {
-        code += "36";
-    } else if color.contains("WHITE") {
-        code += "37";
-    } else {
-        return "".to_string();
-    }
-
-    if color.contains("ON BLACK") {
-        code += ";40";
-    } else if color.contains("ON RED") {
-        code += ";41";
-    } else if color.contains("ON GREEN") {
-        code += ";42";
-    } else if color.contains("ON YELLOW") {
-        code += ";43";
-    } else if color.contains("ON BLUE") {
-        code += ";44";
-    } else if color.contains("ON MAGENTA") {
-        code += ";45";
-    } else if color.contains("ON CYAN") {
-        code += ";46";
-    } else if color.contains("ON WHITE") {
-        code += ";47";
-    }
+    code += bg;
 
     if color.contains("BOLD") {
         code += ";1";
+    }
+
+    if color.contains("DIM") {
+        code += ";2";
     }
 
     if color.contains("ITALIC") {
@@ -146,6 +181,18 @@ pub fn colour(colour_code: &str) -> String {
 
     if color.contains("UNDERLINE") {
         code += ";4";
+    }
+
+    if color.contains("BLINK") {
+        code += ";5";
+    }
+
+    if color.contains("REVERSED") {
+        code += ";7";
+    }
+
+    if color.contains("HIDDEN") {
+        code += ";8";
     }
 
     if color.contains("STRIKETHROUGH") {
@@ -157,7 +204,7 @@ pub fn colour(colour_code: &str) -> String {
 }
 
 /// Add colour escape codes to the given text for printing coloured text in terminal.
-/// This trait is only implemented for `&str` and `String`.
+/// This trait is only implemented for `&str`.
 pub trait Colorizer {
     /// Add colour escape codes to the given text.
     ///
@@ -167,19 +214,19 @@ pub trait Colorizer {
     /// use kdam::term::Colorizer;
     ///
     /// println!("{}", "hello world!".colorize("bold red"));
-    /// println!("{}", "hello world!".colorize("blue on white"));
+    /// println!("{}", "hello world!".colorize("bright white on blue"));
     /// ```
     fn colorize<'a>(&'a self, code: &str) -> String;
 }
 
 impl Colorizer for str {
     fn colorize(&self, code: &str) -> String {
-        colour(code) + self + "\x1b[0m"
-    }
-}
+        let esc_code = colour(code);
 
-impl Colorizer for String {
-    fn colorize(&self, code: &str) -> String {
-        colour(code) + self + "\x1b[0m"
+        if esc_code == "" {
+            self.to_owned()
+        } else {
+            esc_code + self + "\x1b[0m"
+        }
     }
 }